requirements-admin.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
bcrypt==4.0.1
boto3==1.39.3
botocore==1.39.3
certifi==2025.7.9
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.14
fastapi-storages==0.3.0
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
makefun==1.16.0
Mako==1.3.10
MarkupSafe==3.0.2
nest-asyncio==1.6.0
nodeenv==1.9.1
passlib==1.7.4
platformdirs==4.3.8
pre_commit==4.2.0
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
ruff==0.11.11
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.31.2
WTForms==3.1.2
openpyxl==3.1.5
pandas==2.3.1
